test(zqlite): split planner auto flip regression
